V 1.25.2 - patch bug that allowed any user to call the close command

+++ commands/close.js
@@ -4,6 +4,7 @@ module.exports = {
   description: "Closes the channel.",
   command: async function (message) {
     const target = message.channel;
+    console.log(message.channel.name)
     //check for log channel
     const log = message.guild.channels.cache.find(
       (channel) => channel.name === config.log_channel
@@ -11,10 +12,15 @@ module.exports = {
     //check for user permissions
     if (!message.member.hasPermission("MANAGE_CHANNELS")) {
       message.channel.send(`ERROR 401: ${message.author}, Missing permissions.`);
+      return
     }
     if (!log) {
       message.channel.send(`ERROR 404: ${message.author}, Log channel not found.`);
       return;
+    }
+    if (!message.channel.name.includes("suspended")) {
+      message.channel.send(`ERROR 401: Channel cannot be deleted.`);
+      return
     }
     target.delete().catch((e) => console.error(e));
     log.send(`${message.author} deleted a channel.`);
